urdu_moshi/requirements.txt
jax[tpu]>=0.4.25
flax>=0.8.2
optax>=0.2.2
orbax-checkpoint>=0.5.10

transformers>=4.40.0
sentencepiece>=0.2.0
numpy>=1.26.0
librosa>=0.10.1
soundfile>=0.12.1
resampy>=0.4.2

tensorflow>=2.16.0
tensorflow-datasets>=4.9.4

torch>=2.2.0
torchaudio>=2.2.0

whisper-timestamped>=1.15.4
pyannote.audio>=3.1.1

scipy>=1.13.0
einops>=0.8.0
tensorboard>=2.16.0
tqdm>=4.66.0
